tg/requirements.txt
python-telegram-bot==13.15
mysql-connector-python==8.0.32